Pass environment to commands at execution time.
This means we don't have to know what environment a command will be executed
in when we construct it.

The motivation was simplifying subshell parsing: we'll be able to parse a
command fully, including the contents of any subshells, in a single pass.
Previously we couldn't do this because the subshell's environment can't be
created until the command containing the subshell is executed. Look out for
this change in a subsequent commit.

A nice side effect of the change was simplifying the Gitsh::Parser: it no
longer needs a reference to a Gitsh::Environment.