Support alternative names for Readline.
The OpenBSD port for Readline calls the library `ereadline`, which means
that the gitsh configure script needed to be patched to be used on
OpenBSD systems.

This commit introduces a `READLINE_LIB` environment variable, which can be
used to specify an alternative name to be passed to the Ruby C-extension's
configuration script.

Based on Mike Burns' patch to `configure` in the OpenBSD ports tree:
  https://marc.info/?l=openbsd-ports&m=148340820830595&w=2